HS-210 move dto classes into new module
