+++ platform/alarms/rest/pom.xml
@@ -55,6 +55,11 @@
       <groupId>io.swagger.core.v3</groupId>
       <artifactId>swagger-annotations</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.opennms.horizon</groupId>
+      <artifactId>common-model</artifactId>
+      <version>${project.version}</version>
+    </dependency>
   </dependencies>
 
   <build>
