+++ platform/db/common-model/src/main/java/org/opennms/horizon/db/common/model/AlarmAckDTO.java
@@ -1,4 +1,4 @@
-package org.opennms.horizon.alarms.api;
+package org.opennms.horizon.db.common.model;
 
 public class AlarmAckDTO {
     private String user;
