+++ platform/db/common-model/src/main/java/org/opennms/horizon/db/common/model/AlarmDTO.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.db.model.dto;
+package org.opennms.horizon.db.common.model;
 
 import java.util.Date;
 import java.util.List;
