+++ platform/db/common-model/src/main/java/org/opennms/horizon/db/common/model/ReductionKeyMemoDTO.java
@@ -26,21 +26,32 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.db.model.dto;
+package org.opennms.horizon.db.common.model;
 
-import java.util.Collection;
-import java.util.List;
+import java.util.Objects;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+public class ReductionKeyMemoDTO extends MemoDTO {
+    private String reductionKey;
 
-public class AlarmCollectionDTO extends AbstractCollectionDTO<AlarmDTO> {
-    public AlarmCollectionDTO() {
+    public String getReductionKey() {
+        return reductionKey;
     }
-    public AlarmCollectionDTO(final Collection<? extends AlarmDTO> alarms) {
-        objects.addAll(alarms);
+
+    public void setReductionKey(String reductionKey) {
+        this.reductionKey = reductionKey;
     }
-    @JsonProperty("alarms")
-    public List<AlarmDTO> getObjects() {
-        return objects;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        if (!super.equals(o)) return false;
+        ReductionKeyMemoDTO that = (ReductionKeyMemoDTO) o;
+        return Objects.equals(reductionKey, that.reductionKey);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(super.hashCode(), reductionKey);
     }
 }
