+++ platform/db/model/src/main/java/org/opennms/horizon/db/model/mapper/AlarmMapper.java
@@ -41,13 +41,13 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.Mappings;
+import org.opennms.horizon.db.common.model.AlarmDTO;
+import org.opennms.horizon.db.common.model.AlarmSummaryDTO;
+import org.opennms.horizon.db.common.model.EventParameterDTO;
 import org.opennms.horizon.db.model.AckType;
 import org.opennms.horizon.db.model.OnmsAlarm;
 import org.opennms.horizon.db.model.OnmsEventParameter;
 import org.opennms.horizon.db.model.TroubleTicketState;
-import org.opennms.horizon.db.model.dto.AlarmDTO;
-import org.opennms.horizon.db.model.dto.AlarmSummaryDTO;
-import org.opennms.horizon.db.model.dto.EventParameterDTO;
 import org.opennms.horizon.events.api.EventConfDao;
 import org.springframework.beans.factory.annotation.Autowired;
