+++ platform/pom.xml
@@ -85,7 +85,7 @@
         <mina.version>2.1.5</mina.version>
         <mockito.version>3.4.6</mockito.version>
         <netty3.version>3.10.6.Final</netty3.version>
-        <netty4.version>4.1.69.Final</netty4.version>
+        <netty4.version>4.1.78.Final</netty4.version>
         <opennms.tracker.version>0.7</opennms.tracker.version>
         <opentracing.version>0.31.0</opentracing.version>
         <org.osgi.service.jdbc.version>1.0.0</org.osgi.service.jdbc.version>
